api-server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.2.0
numpy==1.26.3
apscheduler==3.10.4
python-dotenv==1.0.1
alembic==1.13.1 